backend/requirements.txt
# Core
pytz==2019.3
Django==3.0.3
django-configurations==2.2
gunicorn==20.0.4
newrelic==5.4.1.134

# For the persistence stores
psycopg2-binary==2.8.4
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.0.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.11.0
Markdown==3.1.1
django-filter==2.2.0

# Developer Tools
ipdb==0.12.3
ipython==7.11.1
mkdocs==1.0.4
flake8==3.7.9

# Testing
mock==3.0.5
factory-boy==2.12.0
django-nose==1.4.6
nose-progressive==1.5.2
coverage==5.0.3

# Static and Media Storage
django-storages==1.8.0
boto3==1.11.1
Pillow==7.0.0
